feat: make all template workflows callable (#124)
# Description

This repository becomes the central repository for all workflows. Thus
we shouldn't copy the workflows to other repositories as this makes it
hard to update them (requires one PR per using repository). Instead we
make the workflows callable from the outside. This way we can manage
everything centrally.

The path for callable workflows is `.github/workflows/` according to the
documentation.

To avoid circular dependencies and creating pull requests again and
again, we disabled the Renovate's auto-merge feature for the workflows
belonging to this repository.

# Verification

Did some checks within the https://github.com/Hapag-Lloyd/test
repository.

# Checklist

- [x] My code follows the style guidelines of the project
- [x] I have performed a self-review of my own code
- [x] I have made corresponding changes to the documentation